Remove unsetvar
The "unsetvar" method is many things.  It is, for example, very, very procedural;
it is also only used in tests.  And in those tests it is only used to show that
it (unsetvar itseslf) "works," or to unset variables that don't need to be unset.
And, last but not least, it is removed by this commit.